Set expel timeout to 5 seconds
Focal has MySQL version 8.0.20-0ubuntu0.20.04.1 which sets
group_replication_member_expel_timeout to 0 by default [0]. As a result,
cluster members are expelled from the cluster 5 seconds after the loss
of connectivity.

This proves to be too short for our CI and some other environments in
the wild.

MySQL documentation for this option suggests that 5 or even 10 seconds
may be too low for congested or overcommitted environments which happens
to be the case. As of MySQL 8.0.21 the default for
group_replication_member_expel_timeout is set to 5 seconds.

[0] https://dev.mysql.com/doc/refman/8.0/en/group-replication-options.html#sysvar_group_replication_member_expel_timeout
[1] https://dev.mysql.com/doc/refman/8.0/en/group-replication-responses-failure-expel.html

Change-Id: Ia33c6263e616657d21b41fcf1a8abfb34e6cb58d
Partial-Bug: #1882205