Handle attempts to write to a RO node
Though the charm took pains to avoid it there are states the charm can
get into in which it attempts to write to a read only node.

This change catches the OperationalError error code 1290: "The MySQL server
is running with the --super-read-only option so it cannot execute this
statement" gracefully.

Change-Id: If9188f5e72701f4bd7575b09217f355fa3d505a2
Closes-Bug: #1882205