Do not RESTART after configure_instance
After the mysql-shell update to 8.0.22 the shell.run_sql("RESTART") in
configure_instance began exiting with non-zero and an error message
stating the MySQL server has gone away.

It seems the shell.run_sql("RESTART") is not necessary. This change
removes this from the configure_instance method. It also adds a
restart_instance method in case we need it in the future.

Closes-Bug: #1901771
Change-Id: I5f5278a4097e75c1a37e1880b594a1e81691249b