Remove instance flags when instance removed
Previously when an instance was removed the leadership settings and
charms.reactive flags remained for that instance's IP address. If a new
instance was subsequently added and happened to have the same IP address
the charm would never add the new instance to the cluster because it
believed the instance was already configured and clustered based on
leader settings.

Clear leader settings flags for instance cluster configured and
clustered.

Due to a bug in Juju the previous use of IP addresses with '.' were
unable to be unset. Transform dotted flags to use '-' instead.

func-test-pr: https://github.com/openstack-charmers/zaza-openstack-tests/pull/565

Change-Id: If3ffa9e9191c057ac7e3d96bfcf84d8a3a2ad45a
Closes-Bug: #1922394
Related-Bug: #1889792