Stop using AUTOMATIC node whitelist
When the group_replication_ip_allowlist *1 is set to the default
value of AUTOMATIC clusters can only be formed on RFC1918 private
networks. This change changes the behavious so that
group_replication_ip_allowlist will now be set to a list of CIDRs
that cover the units that form the cluster.

If a unit is added after the cluster is formed then a new action
is provided to force the charm to update group_replication_ip_allowlist
on all nodes. This is not done automatically as replication needs
to be stopped to update it.

*1 https://dev.mysql.com/doc/refman/8.0/en/group-replication-options.html#sysvar_group_replication_ip_allowlist

Change-Id: I851fee541f85a69e346017d3ec1c12be06efdfd1