clustering: tweak allowlist generation
Instead of trying to resolve the network CIDR from the local unit
for all units in the cluster just use the actual IP addresses of
the cluster unit when generating the IP allowlist for cluster
connectivity.

Also add the network CIDR for the local units cluster address
which is the only one that will be guaranteed to be resolvable.

For deployments where all units are on the same Layer 2 network
addition of units with complete automatically - in Layer 3
routed network topologies new units will be blocked until the
update-unit-acls action is executed which is a service
disruption operation.

Closes-Bug: 1926460
Change-Id: I16e43c37e1af02fb0e23a9c460d70bf5e1dd0fb1