Remove the openstack-upgrade action
The openstack-principle reactive layer brings both the
openstack-origin option and the openstack-upgrade
action, none of which being needed for this charm.
This layer has been used already since the very first
commit of this charm, most likely for no specific
reason.

Also synced tox/pip boilerplate in order to include:
* https://github.com/openstack-charmers/release-tools/pull/156
* https://github.com/openstack-charmers/release-tools/pull/157

Closes-Bug: #1944953
Change-Id: I001f435c648229b65fb718a78a33e6303a0faf11