Update to build using charmcraft - focal batch
The 'focal' batch includes charms that are distro based, rather than
OpenStack release based.

Note: this has a modified charmcraft.yaml to work with bionic / tox
needs to be 3.18 (bionic comes with 2.5), so tox is pip installed as
user in the charmcraft lxd.

The charmcraft build use 18.04 (bionic) to build on, but the charm will
run on 18.04 and 20.04.  This is so that the charm can be used to series
upgrade from bionic to focal.

Change-Id: Id9cb60030ec94ec0494ea23d2016cc68c873e91a