Prometheus mysqld exporter
Install mysqld-exporter with snap. It's a prometheus exporter for mysql,
which provide the http interface for prometheus to pull the metrics.
It should only install & setup configuration when relation connect to prometheus,
and disable when relation remove.
The charm also need to create read only user for the exporter to get the
metrics.

https://github.com/prometheus/mysqld_exporter

Change-Id: Id6a6337004586699c7f943530fa7a32deae26db0
func-test-pr: https://github.com/openstack-charmers/zaza-openstack-tests/pull/859