mysqldump: add --set-gtid-purged=COMMENTED to mysqldump command
The backup file taken from a charm-managed mysql instance contains
"@@GLOBAL.GTID_PURGED" variable at the beginning, which causes
restore operation to fail with the following message: "Cannot
update GTID_PURGED with the Group Replication plugin running".
The workaround for this issue is to supply "--set-gtid-purged=OFF"
to command, which omits the "GTID_PURGED" from the mysqldump output.

This patch fixes the issue by adding "--set-gtid-purged=COMMENTED"
to mysqldump command, which causes  mysqldump to emit the GTID_PURGED
variable as a commented out line. In this way, we don't lose the
GTID_PURGED variable information which may be needed in some scenarios
and can be un-commented out before restoring if needed.

Closes-bug: 1989975
Change-Id: Ie3b71a60e4c7afda48114a53cce9cf713b933d71