Pin tox to < 4.0.0
Tox 4.0.0 was recently released and it has several breaking changes.
We pin to < 4.0.0 here. We are planning to move forward only on the
master charm branches.

Tox is also pinned to < 4.0.0 for stable branches in upstream
openstack-zuul-jobs as well as in zosci-config. However, the
requires= section in the charm's tox.ini file ends up installing
tox>4, wiping out the zuul-pinned tox<4 that was already installed
installed. This patch fixes that.

Additional changes include:
- Add bindep.txt changes from origin/master by hand.
- Switch charm-tools to 2.8.4 to remove ruamel requirement that
  doesn't work on Python 3.10.

Related-Bug: #2002788
Change-Id: If5e5beda90a918e8ff48f1d47fc91d90d79d27fe