Wait for Group Replication to finish; 3100 before commit error
The bug is triggered, as a race, usually by the
prometheus-relation-joined hook, when it tries to create a user whilst
Group Replication is recovering during a rolling restart.  This patch
alters the create_user() method so that it detects the failure condition
and then retries for up to a minute (6 times, every 10 seconds) for the
Group Replication to recover before giving up and returning False
(indicating the the user was not created).  This will usually result in
the handler not completing during the hook, and then retrying on the
next hook.

Change-Id: I5df4fd5ecbdd2b7bce525a9930dcffbc5868cbb8
Closes-Bug: #2018385