Fix status during db-router-relation-departed
The functions get_cluster_instance_mode() and
get_cluster_status_text() should be able to return None if they can't
access the relevant status, but it's possible that keys may exist in the
status and this causes KeyErrors.  This patch makes the functions more
robust such that they return None if the items don't exist.  This only
affects reporting on the status line, rather than functionality in the
charm.

Change-Id: Ie63438cc8801224b0608203f7c935c430ef46045
Closes-Bug: #2020216