Fix side-effect from delayed flag update
Actions such as remove-instance and add-instance
cause leader-settings updates, which in turn result
in flag updates that are later checked in actions
and hooks. However, a hook is required to run after
an action to apply the leader-settings updates into
the flags, causing issues when actions are run
without a hook in-between.

This patch addresses that by changing the flag
checks to leader-settings checks, so it is
always using the most up-to-date information
and preventing the issue.

Closes-bug: #2021880
Change-Id: Ic44fb0895016164f0a3c573c3247443b74e8faa8