Update charmcraft to use reactive plugin/restore 20.04 base
The 20.04 base got removed in a recent commit, but it's still needed
to test the focal bundles as Juju has become more strict about whether a
local charm will install based on charmcraft.yaml bases rather than
metadata (which still indicated that the charm was okay for focal).

Also update the charmcraft.yaml to use the reactive plugin, and convert
the charm to a binary wheel's charm so that it can also support lunar.

Fold in changes to support mantic.

Func-Test-Pr: https://github.com/openstack-charmers/zaza-openstack-tests/pull/1093
Change-Id: I372d4f06e1360a7b8d6f7972a1395118abf6a96a