Delay prom_exporter user creation until password is available.
There are situations where the create_local_prometheus_exporter_user()
handler could be executed in a follower unit before the leader units
gets the chance to generate a password for the prom_exporter user and
store it in the leader databag, this has the consequence that the
prom_exporter user is created with the password 'None' (string
representation of the `None` value), this leads to the problem that the
mysql_exporter daemon can't connect to mysqld.

This patch checks that the password is available, if it isn't, then bail
out and let subsequent executions to re-evaluate this condition.

Closes-Bug: #2033306
Change-Id: I58fc489fd4cc4e06b11c293f63e73bd0ab18897f