Update pinning
The last version of charm-tools is 2.8.7, the pinning is relaxed to
allow it to pick up any future micro or minor releases that fix build
issues.

Migrate the charm from PyYAML 5.3.1 to 6.0.1 to fix fail to build from
source with incompatible Cython.

Add Cython-0.29.36 to src/build.lock so it gets added to the wheelhouse
and allows the building of the python packages during the install hook.

Change-Id: I1590aaa51a63099e4d8c6521dd836ecb19990525