Add gtid-purged-mode option
When creating a backup of a database, the --set-gtid-purged settin must
be set to OFF if the dump is to be restored on a different MySQL
cluster. This change creates a knob in the mysql dump action of the
charm that allows setting this option to one of the supported values.

Closes-Bug: #2040159

Change-Id: Ide1b222c36edbe65de123a57630fe3dc24c4d6b6
Signed-off-by: Gabriel Adrian Samfira <gsamfira@cloudbasesolutions.com>